Silencing compiler warning C4244 (#666)
* Silence some C4244 warnings in MWSE-mwscript with explicit casts

* Change some var types to eliminate more C4244 warnings

* Resolve more C4244 in sound related files.

Add AudioController::getMixVolumeRaw that returns an unsigned char.

* Use explicit casts in DialogueFilterContext

* Use explicit casts when dealing with global variables

* Use explicit casts in math expressions

* Use more explicit casts

* Use different parameter type in tes3alchemy.create

+++ MWSE/ArrayUtil.cpp
@@ -38,8 +38,9 @@ namespace mwse {
 		return value;
 	}
 
-	long Arrays::setValue(std::string_view caller, size_t const id, size_t const index, ArrayItem_t const value) {
-		long success = 0;
+	short Arrays::setValue(std::string_view caller, size_t const id, size_t const index, ArrayItem_t const value) {
+		short success = 0;
+
 		if (id > 0 && id <= arrays.size()) {
 			if (index >= 0) {
 				ContainedArray_t& a = arrays[id - 1];
